Add TestPodGroupAsyncPreemption
